Added soundtest feature.
Added a soundtest feature. It can be enabled by defining
ENABLE_SOUNDTEST when running cmake. For example:
'cmake .. -DENABLE_SOUNDTEST=1'
A new entry will appear in settings and it can be used to access the
soundtest menu. The number represents the sound id, and you can change
the current sound using the arrows. Clicking on 'play' plays the
selected sound. This feature is useful when testing if all sounds are
read, loaded and played correctly, specially useful for checking if
dr_wav works correctly after an update. It also prints the name of the
sound to the console when it plays.
At the time of this commit, there are currently 40 sounds in total
loaded and used by BetterSpades.
If more sounds are added, the soundtest.c file must be modified to add
new entries. Perhaps this could be improved in the future to allow
automatic detection/loading of sounds, without having to manually add
entries to the soundtest.c file.