network.c: Do not send input data for movement when player is airborne
Sending movement player input while the player is airborne is a bad idea.
In example it could cause a player to fly if they were to spam for example key w and jump.
And probably and most likely a whole lot more issues as well.

This simple patch prevents this and does not send the packet if player is airborne (In air)