src: Add restock flag to weapon_set
When restocking we do not want to set
the ammo of the player to max.
Player has to do that manually and reload.

This would give player 1 more full clip
and skip a restock wait time in case of no
ammo thus lets fix this so we dont
give any player any advantage